iox-#1099 Move ErrorHandlerMock to mock directory
