+++ iceoryx_hoofs/test/moduletests/test_cxx_functional_interface_expect.cpp
@@ -15,7 +15,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 #include "iceoryx_hoofs/error_handling/error_handling.hpp"
-#include "iceoryx_hoofs/testing/error_handler_mock.hpp"
+#include "iceoryx_hoofs/testing/mocks/error_handler_mock.hpp"
 
 #include "test_cxx_functional_interface_types.hpp"
