+++ iceoryx_posh/test/test.hpp
@@ -16,7 +16,7 @@
 #ifndef IOX_POSH_TEST_TEST_HPP
 #define IOX_POSH_TEST_TEST_HPP
 
-#include "iceoryx_hoofs/testing/error_handler_mock.hpp"
+#include "iceoryx_hoofs/testing/mocks/error_handler_mock.hpp"
 
 #include <fstream>
 #include <gmock/gmock.h>
