+++ iceoryx_posh/testing/roudi_environment/roudi_environment.cpp
@@ -19,7 +19,7 @@
 #include "iceoryx_hoofs/cxx/helplets.hpp"
 #include "iceoryx_hoofs/internal/posix_wrapper/shared_memory_object/memory_map.hpp"
 #include "iceoryx_hoofs/log/logmanager.hpp"
-#include "iceoryx_hoofs/testing/error_handler_mock.hpp"
+#include "iceoryx_hoofs/testing/mocks/error_handler_mock.hpp"
 #include "iceoryx_posh/internal/popo/building_blocks/unique_port_id.hpp"
 #include "iceoryx_posh/internal/roudi/roudi.hpp"
 #include "iceoryx_posh/runtime/posh_runtime.hpp"
